Update ColorSlider with updates from Icon-O-Matic
Update ColorSlider so that you can hold the mouse down and slide
to change the color continuously instead of it updating on MouseUp.
It also updates much faster.

Rearrange the methods to a nice order. Virtual method overrides first,
then public members in pseudo-alphabetical order, then private methods.

Rename private member methods to begin with an _ as in other classes.